@@ -30,104 +36,103 @@
         ]
     },
     "dependencies": {
-        "antd": "^4.19.1",
-        "dayjs": "^1.10.8",
-        "jquery": "^3.6.0",
-        "lodash": "^4.17.21",
-        "normalize.css": "^8.0.1",
-        "react": "17.0.2",
-        "react-dom": "17.0.2",
-        "react-router-dom": "^6.2.2"
+        "antd": "5.0.1",
+        "dayjs": "1.11.6",
+        "jquery": "3.6.1",
+        "lodash": "4.17.21",
+        "normalize.css": "8.0.1",
+        "react": "18.2.0",
+        "react-dom": "18.2.0",
+        "react-router-dom": "6.4.3",
+        "webextension-polyfill": "0.10.0"
     },
     "devDependencies": {
-        "@babel/core": "^7.17.5",
-        "@babel/plugin-proposal-decorators": "^7.17.2",
-        "@babel/plugin-transform-react-constant-elements": "^7.17.6",
-        "@babel/plugin-transform-runtime": "^7.17.0",
-        "@babel/preset-env": "^7.16.11",
-        "@babel/preset-react": "^7.16.7",
-        "@babel/preset-typescript": "^7.16.7",
-        "@lukeapage/webpack-hot-middleware": "2.25.2",
-        "@nuxt/friendly-errors-webpack-plugin": "^2.5.2",
-        "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
-        "@types/chrome": "0.0.179",
-        "@types/cors": "^2.8.12",
-        "@types/express": "^4.17.13",
-        "@types/html-minifier": "^4.0.2",
-        "@types/html-webpack-plugin": "^3.2.6",
-        "@types/jquery": "^3.5.14",
-        "@types/lodash": "^4.14.179",
-        "@types/node": "^17.0.21",
-        "@types/react": "^17.0.40",
-        "@types/react-dom": "^17.0.13",
-        "@types/react-router-dom": "^5.3.3",
-        "@types/webpack": "^5.28.0",
-        "@types/webpack-bundle-analyzer": "^4.4.1",
-        "@types/webpack-hot-middleware": "^2.25.6",
-        "@types/webpackbar": "^4.0.3",
-        "@types/yargs": "^17.0.9",
-        "@typescript-eslint/eslint-plugin": "^5.14.0",
-        "@typescript-eslint/parser": "^5.14.0",
-        "antd-dayjs-webpack-plugin": "^1.0.6",
-        "audit-ci": "6.0.0-beta.1",
-        "babel-loader": "^8.2.3",
-        "babel-plugin-import": "^1.13.3",
-        "babel-plugin-lodash": "^3.3.4",
-        "chalk": "^4.1.2",
-        "clean-webpack-plugin": "^4.0.0",
-        "copy-webpack-plugin": "^10.2.4",
-        "core-js": "^3.21.1",
-        "cors": "^2.8.5",
-        "cross-env": "^7.0.3",
-        "css-loader": "^6.7.1",
-        "css-minimizer-webpack-plugin": "^3.4.1",
-        "eslint": "^8.11.0",
-        "eslint-config-airbnb": "^19.0.4",
-        "eslint-config-prettier": "^8.5.0",
-        "eslint-import-resolver-typescript": "^2.5.0",
-        "eslint-plugin-eslint-comments": "^3.2.0",
-        "eslint-plugin-import": "^2.25.4",
-        "eslint-plugin-jsx-a11y": "^6.5.1",
-        "eslint-plugin-promise": "^6.0.0",
-        "eslint-plugin-react": "^7.29.3",
-        "eslint-plugin-react-hooks": "^4.3.0",
-        "eslint-plugin-unicorn": "^41.0.0",
-        "execa": "^5.1.1",
-        "express": "^4.17.3",
-        "fork-ts-checker-webpack-plugin": "6.5.0",
-        "format-imports": "^3.0.1",
-        "get-port": "^5.1.1",
-        "html-webpack-plugin": "^5.5.0",
-        "http-proxy-middleware": "^2.0.3",
-        "husky": "^7.0.4",
-        "less": "^4.1.2",
-        "less-loader": "^10.2.0",
-        "lint-staged": "^12.3.5",
-        "mini-css-extract-plugin": "^2.6.0",
-        "open": "^8.4.0",
-        "postcss-less": "^6.0.0",
-        "postcss-scss": "^4.0.3",
-        "prettier": "^2.5.1",
-        "react-devtools": "4.24.0",
-        "react-refresh": "^0.11.0",
-        "rimraf": "^3.0.2",
-        "sass": "^1.49.9",
-        "sass-loader": "^12.6.0",
-        "ssestream": "^1.1.0",
-        "stylelint": "^14.5.3",
-        "stylelint-config-prettier": "^9.0.3",
-        "stylelint-config-recess-order": "^3.0.0",
-        "stylelint-config-standard": "^25.0.0",
-        "stylelint-declaration-block-no-ignored-properties": "^2.5.0",
-        "stylelint-scss": "^4.1.0",
-        "terser-webpack-plugin": "^5.3.1",
-        "ts-node": "^10.7.0",
-        "typescript": "^4.6.2",
-        "webpack": "^5.70.0",
-        "webpack-bundle-analyzer": "^4.5.0",
-        "webpack-dev-middleware": "^5.3.1",
-        "webpack-merge": "^5.8.0",
-        "webpackbar": "^5.0.2",
-        "yargs": "^17.3.1"
+        "@babel/core": "7.20.2",
+        "@babel/plugin-proposal-decorators": "7.20.2",
+        "@babel/plugin-transform-react-constant-elements": "7.20.2",
+        "@babel/plugin-transform-runtime": "7.19.6",
+        "@babel/preset-env": "7.20.2",
+        "@babel/preset-react": "7.18.6",
+        "@babel/preset-typescript": "7.18.6",
+        "@nuxt/friendly-errors-webpack-plugin": "2.5.2",
+        "@pmmmwh/react-refresh-webpack-plugin": "0.5.10",
+        "@swc/core": "1.3.19",
+        "@types/chrome": "0.0.203",
+        "@types/cors": "2.8.12",
+        "@types/express": "4.17.14",
+        "@types/html-minifier": "4.0.2",
+        "@types/html-webpack-plugin": "3.2.6",
+        "@types/jquery": "3.5.14",
+        "@types/lodash": "4.14.190",
+        "@types/node": "18.11.9",
+        "@types/react": "18.0.25",
+        "@types/react-dom": "18.0.9",
+        "@types/react-router-dom": "5.3.3",
+        "@types/wait-on": "5.3.1",
+        "@types/webextension-polyfill": "0.9.1",
+        "@types/webpack": "5.28.0",
+        "@types/webpack-bundle-analyzer": "4.6.0",
+        "@types/webpack-hot-middleware": "2.25.6",
+        "@types/webpackbar": "4.0.3",
+        "@types/yargs": "17.0.14",
+        "@yutengjing/eslint-config-react": "0.0.8",
+        "@yutengjing/prettier-config": "0.0.1",
+        "ansi-colors": "4.1.3",
+        "antd-dayjs-webpack-plugin": "1.0.6",
+        "babel-loader": "9.1.0",
+        "babel-plugin-import": "1.13.5",
+        "babel-plugin-lodash": "3.3.4",
+        "browserslist": "4.21.4",
+        "chokidar": "3.5.3",
+        "consola": "2.15.3",
+        "copy-webpack-plugin": "11.0.0",
+        "core-js": "3.26.1",
+        "cors": "2.8.5",
+        "cross-env": "7.0.3",
+        "css-loader": "6.7.2",
+        "css-minimizer-webpack-plugin": "4.2.2",
+        "eslint": "8.28.0",
+        "exit-hook": "2.2.1",
+        "express": "4.18.2",
+        "fork-ts-checker-webpack-plugin": "7.2.13",
+        "format-imports": "3.1.3",
+        "get-port": "5.1.1",
+        "html-webpack-plugin": "5.5.0",
+        "html-webpack-tags-plugin": "3.0.2",
+        "http-proxy-middleware": "2.0.6",
+        "husky": "8.0.2",
+        "less": "4.1.3",
+        "less-loader": "11.1.0",
+        "lightningcss": "1.16.1",
+        "lint-staged": "13.0.4",
+        "mini-css-extract-plugin": "2.7.0",
+        "open": "8.4.0",
+        "postcss-less": "6.0.0",
+        "postcss-scss": "4.0.5",
+        "prettier": "2.8.0",
+        "react-devtools": "4.26.1",
+        "react-refresh": "0.14.0",
+        "rimraf": "3.0.2",
+        "sass": "1.56.1",
+        "sass-loader": "13.2.0",
+        "ssestream": "1.1.0",
+        "stylelint": "14.15.0",
+        "stylelint-config-prettier": "9.0.4",
+        "stylelint-config-recess-order": "3.0.0",
+        "stylelint-config-standard": "29.0.0",
+        "stylelint-declaration-block-no-ignored-properties": "2.6.0",
+        "stylelint-scss": "4.3.0",
+        "terser-webpack-plugin": "5.3.6",
+        "ts-node": "10.9.1",
+        "typescript": "4.9.3",
+        "wait-on": "6.0.1",
+        "webext-bridge": "5.0.5",
+        "webpack": "5.75.0",
+        "webpack-bundle-analyzer": "4.7.0",
+        "webpack-dev-middleware": "6.0.0",
+        "webpack-hot-middleware": "2.25.3",
+        "webpack-merge": "5.8.0",
+        "webpackbar": "5.0.2",
+        "yargs": "17.6.2"
     }
 }
