@@ -1,12 +1,14 @@
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { HashRouter } from 'react-router-dom';
 
 import App from './App';
+
 import './App.scss';
 
-ReactDOM.render(
+const container = document.querySelector('#root');
+const root = createRoot(container!);
+root.render(
     <HashRouter>
         <App />
     </HashRouter>,
-    document.querySelector('#root'),
 );
