@@ -1,17 +1,13 @@
-/* eslint-disable camelcase, import/no-dynamic-require, global-require, @typescript-eslint/no-var-requires */
 module.exports = (api) => {
-    const isProd = api.env('production');
-    const { minimum_chrome_version } = require(`./src/manifest.${isProd ? 'prod' : 'dev'}.json`);
+    api.cache(true);
+
     const envPreset = [
         '@babel/env',
         {
             modules: false,
-            targets: minimum_chrome_version
-                ? `Chrome > ${minimum_chrome_version}`
-                : 'last 2 Chrome versions',
             bugfixes: true,
             useBuiltIns: 'usage',
-            corejs: { version: require('./package.json').devDependencies['core-js'].slice(1) },
+            corejs: { version: require('./package.json').devDependencies['core-js'] },
         },
     ];
 
@@ -38,7 +34,7 @@ module.exports = (api) => {
                 plugins: [require.resolve('react-refresh/babel')],
             },
             production: {
-                presets: ['@babel/preset-react'],
+                presets: [['@babel/preset-react', { runtime: 'automatic', development: false }]],
                 plugins: ['@babel/plugin-transform-react-constant-elements'],
             },
         },
