@@ -1,4 +1,4 @@
-import { Options } from 'http-proxy-middleware/dist/types';
+import type { Options } from 'http-proxy-middleware/dist/types';
 
 export interface ProxyTable {
     [path: string]: Options;
