@@ -1,13 +1,16 @@
 import AntdDayjsWebpackPlugin from 'antd-dayjs-webpack-plugin';
+import browserslist from 'browserslist';
 import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
 import ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
-import { resolve } from 'path';
+import lightningCss from 'lightningcss';
 import TerserPlugin from 'terser-webpack-plugin';
 import webpack, { BannerPlugin } from 'webpack';
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
 import merge from 'webpack-merge';
 
-import { COPYRIGHT, ENABLE_ANALYZE, PROJECT_ROOT } from '../utils/constants';
+import pkg from '../../package.json';
+import { __DEV__, COPYRIGHT, ENABLE_ANALYZE } from '../utils/constants';
+import { resolveSrc } from '../utils/path';
 import commonConfig from './webpack.common';
 
 const prodConfig = merge(commonConfig, {
@@ -20,7 +23,7 @@ const prodConfig = merge(commonConfig, {
         new ForkTsCheckerWebpackPlugin({
             typescript: {
                 memoryLimit: 1024 * 2,
-                configFile: resolve(PROJECT_ROOT, 'src/tsconfig.json'),
+                configFile: resolveSrc('tsconfig.json'),
                 profile: ENABLE_ANALYZE,
             },
         }),
@@ -30,7 +33,6 @@ const prodConfig = merge(commonConfig, {
             hashDigestLength: 20,
         }),
         new AntdDayjsWebpackPlugin(),
-        new CssMinimizerPlugin(),
     ],
     optimization: {
         splitChunks: {
@@ -45,8 +47,22 @@ const prodConfig = merge(commonConfig, {
         minimize: true,
         minimizer: [
             new TerserPlugin({
+                minify: TerserPlugin.swcMinify,
                 parallel: true,
                 extractComments: false,
+            }),
+            new CssMinimizerPlugin({
+                minify: CssMinimizerPlugin.lightningCssMinify,
+                minimizerOptions: {
+                    // @ts-expect-error webpack type define wrong
+                    targets: lightningCss.browserslistToTargets(browserslist(pkg.browserslist)),
+                    preset: [
+                        'default',
+                        {
+                            discardComments: { removeAll: true },
+                        },
+                    ],
+                },
             }),
         ],
     },
