@@ -1,5 +1,6 @@
-import { Express } from 'express';
-import chalk from 'chalk';
+import chalk from 'ansi-colors';
+import console from 'consola';
+import type { Express } from 'express';
 import { createProxyMiddleware } from 'http-proxy-middleware';
 
 import proxyTable from '../configs/proxy';
@@ -12,9 +13,8 @@ export default function proxyMiddleware(server: Express): void {
     Object.entries(proxyTable).forEach(([path, options]) => {
         const from = path;
         const to = options.target as string;
-        console.log(`proxy ${link(from)} ${chalk.green('->')} ${link(to)}`);
+        console.info(`proxy ${link(from)} ${chalk.green('->')} ${link(to)}`);
 
-        // eslint-disable-next-line no-param-reassign
         if (!options.logLevel) options.logLevel = 'warn';
         server.use(path, createProxyMiddleware(options));
     });
