@@ -1,11 +1,10 @@
-import webpackHotMiddleware from '@lukeapage/webpack-hot-middleware';
-import { Compiler } from 'webpack';
+import type { Compiler } from 'webpack';
 import webpackDevMiddleware from 'webpack-dev-middleware';
+import webpackHotMiddleware from 'webpack-hot-middleware';
 
 import devConfig from '../configs/webpack.dev';
 import { HRM_PATH } from '../utils/constants';
 
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export default function (compiler: Compiler) {
     const publicPath = devConfig.output!.publicPath! as string;
     return [
