@@ -1,20 +1,18 @@
-import execa from 'execa';
-import fs from 'fs';
-import { resolve } from 'path';
+import fs from 'node:fs';
 
-import { __DEV__, ENABLE_DEVTOOLS, HOST, HRM_PATH, PORT } from './constants';
+import { __DEV__, HOST, HRM_PATH, PORT } from './constants';
+import { resolveServer, resolveSrc } from './path';
 
-const src = resolve(__dirname, '../../src');
 const HMR_URL = encodeURIComponent(`http://${HOST}:${PORT}${HRM_PATH}`);
 // !: 必须指定 path 为 devServer 的地址，不然的话热更新 client 会向 chrome://xxx 请求
-const HMRClientScript = `@lukeapage/webpack-hot-middleware/client?path=${HMR_URL}&reload=true&overlay=true`;
+const HMRClientScript = `webpack-hot-middleware/client?path=${HMR_URL}&reload=true&overlay=true`;
 
-const backgroundPath = resolve(src, './background/index.ts');
-const optionsPath = resolve(src, './options/index.tsx');
-const popupPath = resolve(src, './popup/index.tsx');
+const backgroundPath = resolveSrc('background/index.ts');
+const optionsPath = resolveSrc('options/index.tsx');
+const popupPath = resolveSrc('popup/index.tsx');
 
 const devEntry: Record<string, string[]> = {
-    background: [HMRClientScript, backgroundPath],
+    background: [backgroundPath],
     options: [HMRClientScript, optionsPath],
     popup: [HMRClientScript, popupPath],
 };
@@ -25,20 +23,11 @@ const prodEntry: Record<string, string[]> = {
 };
 const entry = __DEV__ ? devEntry : prodEntry;
 
-if (ENABLE_DEVTOOLS) {
-    entry.options.unshift('react-devtools');
-    entry.popup.unshift('react-devtools');
-    execa.command('npx react-devtools').catch((error) => {
-        console.error('Startup react-devtools occur error');
-        error && console.error(error);
-    });
-}
-
-const contentsDirs = fs.readdirSync(resolve(src, 'contents'));
+const contentsDirs = fs.readdirSync(resolveSrc('contents'));
 const validExtensions = ['tsx', 'ts'];
 contentsDirs.forEach((contentScriptDir) => {
     const hasValid = validExtensions.some((ext) => {
-        const abs = resolve(src, `contents/${contentScriptDir}/index.${ext}`);
+        const abs = resolveSrc(`contents/${contentScriptDir}/index.${ext}`);
         if (fs.existsSync(abs)) {
             entry[contentScriptDir] = [abs];
             return true;
@@ -48,15 +37,15 @@ contentsDirs.forEach((contentScriptDir) => {
     });
 
     if (!hasValid) {
-        const dir = resolve(src, `contents/${contentScriptDir}`);
+        const dir = resolveSrc(`contents/${contentScriptDir}`);
         throw new Error(`You must put index.tsx or index.ts under directory: ${dir}`);
     }
 });
 
 // NOTE: 有可能用户没打算开发 content script，所以 contents/all 这个文件夹可能不存在
 if (entry.all && __DEV__) {
-    entry.all.unshift(resolve(__dirname, './allTabClient.ts'));
-    entry.background.unshift(resolve(__dirname, './backgroundClient.ts'));
+    entry.all.unshift(resolveServer('client/allTabClient.ts'));
+    entry.background.unshift(resolveServer('client/backgroundClient.ts'));
 }
 
 export default entry;
